fix: moving the place order logic into webhook processing part
